refactor(api): move findPaginatedFilteredSupParticipant to learner-list bounded context
