+++ api/lib/application/organizations/index.js
@@ -751,39 +751,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/organizations/{id}/sup-participants',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents,
-            assign: 'belongsToSupOrganizationAndManagesStudents',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-          query: Joi.object({
-            'page[size]': Joi.number().integer().empty(''),
-            'page[number]': Joi.number().integer().empty(''),
-            'filter[certificability][]': [Joi.string(), Joi.array().items(Joi.string())],
-            'filter[groups][]': [Joi.string(), Joi.array().items(Joi.string())],
-            'filter[search]': Joi.string().empty(''),
-            'filter[studentNumber]': Joi.string().empty(''),
-            'sort[participationCount]': Joi.string().empty(''),
-            'sort[lastnameSort]': Joi.string().empty(''),
-          }),
-        },
-        handler: organizationController.findPaginatedFilteredSupParticipants,
-        tags: ['api', 'organization', 'sup-participants'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés membres d'un espace Orga**\n" +
-            '- Récupération des étudiants liés à une organisation SUP\n',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/organizations/{id}/sco-participants',
       config: {
         pre: [
