+++ api/server.js
@@ -21,13 +21,14 @@ import {
   complementaryCertificationRoutes,
 } from './src/certification/complementary-certification/routes.js';
 import { learnerManagementRoutes } from './src/prescription/learner-management/routes.js';
+import { learnerListRoutes } from './src/prescription/learner-list/routes.js';
 import { prescriberManagementRoutes } from './src/shared/prescriber-management/routes.js';
 import { devcompRoutes } from './src/devcomp/routes.js';
 import { schoolRoutes } from './src/school/routes.js';
 import { scenarioSimulatorRoutes } from './src/certification/flash-certification/routes.js';
 
 const certificationRoutes = [certificationSessionRoutes, attachTargetProfileRoutes, complementaryCertificationRoutes];
-const prescriptionRoutes = [learnerManagementRoutes];
+const prescriptionRoutes = [learnerManagementRoutes, learnerListRoutes];
 const sharedRoutes = [prescriberManagementRoutes];
 
 monitoringTools.installHapiHook();
