+++ api/src/prescription/learner-list/domain/usescases/index.js
@@ -1,0 +1,24 @@
+import { fileURLToPath } from 'node:url';
+import { dirname, join } from 'node:path';
+
+import * as supOrganizationParticipantRepository from '../../infrastructure/repositories/sup-organization-participant-repository.js';
+
+import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
+
+const dependencies = {
+  supOrganizationParticipantRepository,
+};
+
+const path = dirname(fileURLToPath(import.meta.url));
+
+const usecasesWithoutInjectedDependencies = {
+  ...(await importNamedExportsFromDirectory({
+    path: join(path, './'),
+    ignoredFileNames: ['index.js'],
+  })),
+};
+
+const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
+
+export { usecases };
