+++ api/src/prescription/learner-list/routes.js
@@ -1,0 +1,5 @@
+import * as supLearnerListRoute from './application/sup-learner-list-route.js';
+
+const learnerListRoutes = [supLearnerListRoute];
+
+export { learnerListRoutes };
