+++ api/tests/integration/application/organizations/index_test.js
@@ -319,77 +319,6 @@ describe('Integration | Application | Organizations | Routes', function () {
     });
   });
 
-  describe('GET /api/organizations/:id/sup-participants', function () {
-    it('should call the organization controller to return sup participants', async function () {
-      // given
-      const method = 'GET';
-      const url = '/api/organizations/1/sup-participants';
-
-      sinon
-        .stub(securityPreHandlers, 'checkUserBelongsToSupOrganizationAndManagesStudents')
-        .callsFake((request, h) => h.response(true));
-      sinon
-        .stub(organizationController, 'findPaginatedFilteredSupParticipants')
-        .callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request(method, url);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      expect(organizationController.findPaginatedFilteredSupParticipants).to.have.been.calledOnce;
-    });
-
-    describe('When parameters are not valid', function () {
-      it('should throw an error when page size is invalid', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/1/sup-participants?page[size]=blabla';
-
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-
-      it('should throw an error when page number is invalid', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/1/sup-participants?page[number]=blabla';
-
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-
-      it('should throw an error when id is invalid', async function () {
-        // given
-        const method = 'GET';
-        const url = '/api/organizations/wrongId/sup-participants';
-
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(method, url);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-    });
-  });
-
   describe('GET /api/organizations/{id}/participants', function () {
     const method = 'GET';
     const url = '/api/organizations/1/participants';
