+++ api/tests/integration/application/organizations/organization-controller_test.js
@@ -1,9 +1,9 @@
 import { expect, sinon, domainBuilder, HttpTestServer } from '../../../test-helper.js';
 import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
+
 import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
 import { ScoOrganizationParticipant } from '../../../../lib/domain/read-models/ScoOrganizationParticipant.js';
-import { SupOrganizationParticipant } from '../../../../lib/domain/read-models/SupOrganizationParticipant.js';
 import { getCertificationAttestationsPdf as certificationAttestationPdf } from '../../../../lib/infrastructure/utils/pdf/certification-attestation-pdf.js';
 import * as moduleUnderTest from '../../../../lib/application/organizations/index.js';
 import { NoCertificationAttestationForDivisionError } from '../../../../lib/domain/errors.js';
@@ -16,7 +16,6 @@ describe('Integration | Application | Organizations | organization-controller', 
     sandbox = sinon.createSandbox();
     sandbox.stub(usecases, 'findPaginatedFilteredOrganizationMemberships');
     sandbox.stub(usecases, 'findPaginatedFilteredScoParticipants');
-    sandbox.stub(usecases, 'findPaginatedFilteredSupParticipants');
     sandbox.stub(usecases, 'createOrganizationInvitations');
     sandbox.stub(usecases, 'acceptOrganizationInvitation');
     sandbox.stub(usecases, 'findPendingOrganizationInvitations');
@@ -164,53 +163,6 @@ describe('Integration | Application | Organizations | organization-controller', 
 
           // when
           const response = await httpTestServer.request('GET', '/api/organizations/1234/sco-participants');
-
-          // then
-          expect(response.statusCode).to.equal(403);
-        });
-      });
-    });
-  });
-
-  describe('#findPaginatedFilteredSupParticipants', function () {
-    context('Success cases', function () {
-      it('should return an HTTP response with status code 200', async function () {
-        // given
-        const supOrganizationParticipant = new SupOrganizationParticipant();
-        usecases.findPaginatedFilteredSupParticipants.resolves({ data: [supOrganizationParticipant] });
-        securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents.returns(true);
-
-        // when
-        const response = await httpTestServer.request('GET', '/api/organizations/1234/sup-participants');
-
-        // then
-        expect(response.statusCode).to.equal(200);
-      });
-
-      it('should return an HTTP response formatted as JSON:API', async function () {
-        // given
-        const supOrganizationParticipant = new SupOrganizationParticipant();
-        usecases.findPaginatedFilteredSupParticipants.resolves({ data: [supOrganizationParticipant] });
-        securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents.returns(true);
-
-        // when
-        const response = await httpTestServer.request('GET', '/api/organizations/1234/sup-participants');
-
-        // then
-        expect(response.result.data[0].type).to.equal('sup-organization-participants');
-      });
-    });
-
-    context('Error cases', function () {
-      context('when user is not allowed to access resource', function () {
-        it('should resolve a 403 HTTP response', async function () {
-          // given
-          securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents.callsFake((request, h) => {
-            return Promise.resolve(h.response().code(403).takeover());
-          });
-
-          // when
-          const response = await httpTestServer.request('GET', '/api/organizations/1234/sup-participants');
 
           // then
           expect(response.statusCode).to.equal(403);
