+++ api/tests/prescription/learner-list/unit/domain/usecases/find-paginated-filtered-sup-participants_test.js
@@ -1,5 +1,5 @@
-import { expect, sinon } from '../../../test-helper.js';
-import { findPaginatedFilteredSupParticipants } from '../../../../lib/domain/usecases/find-paginated-filtered-sup-participants.js';
+import { expect, sinon } from '../../../../../test-helper.js';
+import { findPaginatedFilteredSupParticipants } from '../../../../../../src/prescription/learner-list/domain/usescases/find-paginated-filtered-sup-participants.js';
 
 describe('Unit | UseCase | findPaginatedFilteredSupParticipants', function () {
   let supOrganizationParticipantRepository;
