+++ api/tests/unit/infrastructure/serializers/jsonapi/organization/sup-organization-participants-serializer_test.js
@@ -1,5 +1,5 @@
 import { expect } from '../../../../../test-helper.js';
-import { SupOrganizationParticipant } from '../../../../../../lib/domain/read-models/SupOrganizationParticipant.js';
+import { SupOrganizationParticipant } from '../../../../../../src/prescription/learner-list/domain/read-models/SupOrganizationParticipant.js';
 import * as serializer from '../../../../../../lib/infrastructure/serializers/jsonapi/organization/sup-organization-participants-serializer.js';
 import { CampaignParticipationStatuses as campaignParticipationsStatuses } from '../../../../../../lib/domain/models/CampaignParticipationStatuses.js';
